// common/MISA.QLCH.CommonStores/src/models/voucher/IVoucher.ts
export default interface IVoucher {

    PublisherVoucher?: string;
    NameVoucher?: string;
    DayBeginVoucher?: Date;
    DayEndVoucher?: Date;
    DescribeVoucher?: string;
    PriceVoucher?: number;
    TotalNumberVoucher?: number;
    TotalPriceVoucher?: number;
    StatusVoucher?: string;

}